MenuController

+++ eRestaurant Demo/eRestaurant.Framework/Entities/Item.cs
@@ -12,8 +12,8 @@ namespace eRestaurant.Framework.Entities
         [Key]
         public int ItemID { get; set; }
         public string Description { get; set; }
-        public double CurrentPrice { get; set; }
-        public double CurrentCost { get; set; }
+        public decimal CurrentPrice { get; set; }
+        public decimal CurrentCost { get; set; }
         public bool Active { get; set; }
         public int? Calories { get; set; }
         public string Comment { get; set; }
